+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### MINOR CHANGES
 * The `boundary` argument can be specified using `bbox` objects. The `bbox` object is converted to `sfc` object with `sf::st_as_sfc` and preserves the same CRS. 
+* Added a more informative error message when `oe_get()` or `oe_read()` are run with empty or unnamed arguments in `...` (#234).
 
 ### DOCUMENTATION FIXES
 * Update description for `boundary` and `boundary_type` arguments.